Requiring the use of wine in order to generate the .paks to the appropriate binary size
There was an issue seen when attempting to use UnrealPak 5.4.2 and 5.6
where the generated binaries were much larger despite the
-compress/-compressed flag. There is not an answer as to why that is the
case at this point in time.

As a result, there is a requirement for the end-user to install wine so
that the .paks can be generated properly. Do note that the UnrealPak
provided by the original Repo is 4.17.2. This was discovered by looking
at the details of the UnrealPak.target.